Rename new plugin to ceremonies-debtor-list
