+++ strapi/src/plugins/ceremonies-debtor-list/admin/src/components/Initializer/index.tsx
