+++ strapi/src/plugins/ceremonies-debtor-list/admin/src/components/PluginIcon/index.tsx
