+++ strapi/src/plugins/ceremonies-debtor-list/admin/src/index.tsx
