+++ strapi/src/plugins/ceremonies-debtor-list/admin/src/pages/App/index.tsx
