+++ strapi/src/plugins/ceremonies-debtor-list/admin/src/pluginId.ts
