+++ strapi/src/plugins/ceremonies-debtor-list/custom.d.ts
