+++ strapi/src/plugins/ceremonies-debtor-list/package.json
