+++ strapi/src/plugins/ceremonies-debtor-list/remove-ceremonies-cron-task.ts
@@ -1,5 +1,5 @@
-import moment from "moment/moment";
-import "moment-timezone";
+import moment from "moment/moment"
+import "moment-timezone"
 
 // Marianum says they cannot show ceremonies older than 3 days, because they might not have consent for it.
 // Filtering them while querying is not enough as it would be possible to query them with custom query.
@@ -8,7 +8,7 @@ export const removeCeremoniesTask = {
     const olderThanThreeDays = moment
       .tz("Europe/Bratislava")
       .subtract("4", "days")
-      .endOf("day");
+      .endOf("day")
 
     await strapi.db.query("api::ceremony.ceremony").deleteMany({
       where: {
@@ -16,16 +16,16 @@ export const removeCeremoniesTask = {
           $lte: olderThanThreeDays.toISOString(),
         },
       },
-    });
+    })
 
-    const meilisearch = strapi.plugin("meilisearch").service("meilisearch");
+    const meilisearch = strapi.plugin("meilisearch").service("meilisearch")
 
     await meilisearch.updateContentTypeInMeiliSearch({
       contentType: "api::ceremony.ceremony",
-    });
+    })
   },
   options: {
     rule: "0 0,12 * * *",
     tz: "Europe/Bratislava",
   },
-};
+}
