+++ strapi/src/plugins/ceremonies-debtor-list/server/bootstrap.ts
