+++ strapi/src/plugins/ceremonies-debtor-list/server/config/index.ts
