+++ strapi/src/plugins/ceremonies-debtor-list/server/content-types/index.ts
