+++ strapi/src/plugins/ceremonies-debtor-list/server/controllers/index.ts
