+++ strapi/src/plugins/ceremonies-debtor-list/server/destroy.ts
