+++ strapi/src/plugins/ceremonies-debtor-list/server/middlewares/index.ts
