+++ strapi/src/plugins/ceremonies-debtor-list/server/register.ts
