+++ strapi/src/plugins/ceremonies-debtor-list/server/routes/index.ts
