+++ strapi/src/plugins/ceremonies-debtor-list/strapi-admin.js
