+++ strapi/src/plugins/ceremonies-debtor-list/tsconfig.json
